pyproject.toml
[project]
name = "connect-deployment-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "streamlit>=1.46.1",
]
